drop look subcommand
