+++ cmd_get.go
@@ -1,8 +1,16 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"sync"
 
+	"github.com/motemen/ghq/cmdutil"
 	"github.com/urfave/cli/v2"
 )
 
@@ -29,7 +37,70 @@ func doGet(c *cli.Context) error {
 		return err
 	}
 	if andLook {
-		return doLook(c)
+		return look(argURL)
 	}
 	return nil
+}
+
+func detectShell() string {
+	shell := os.Getenv("SHELL")
+	if shell != "" {
+		return shell
+	}
+	if runtime.GOOS == "windows" {
+		return os.Getenv("COMSPEC")
+	}
+	return "/bin/sh"
+}
+
+func look(name string) error {
+	var (
+		reposFound []*LocalRepository
+		mu         sync.Mutex
+	)
+	if err := walkAllLocalRepositories(func(repo *LocalRepository) {
+		if repo.Matches(name) {
+			mu.Lock()
+			reposFound = append(reposFound, repo)
+			mu.Unlock()
+		}
+	}); err != nil {
+		return err
+	}
+
+	if len(reposFound) == 0 {
+		if url, err := newURL(name, false, false); err == nil {
+			repo, err := LocalRepositoryFromURL(url)
+			if err != nil {
+				return err
+			}
+			_, err = os.Stat(repo.FullPath)
+
+			// if the directory exists
+			if err == nil {
+				reposFound = append(reposFound, repo)
+			}
+		}
+	}
+
+	switch len(reposFound) {
+	case 0:
+		return fmt.Errorf("No repository found")
+	case 1:
+		repo := reposFound[0]
+		cmd := exec.Command(detectShell())
+		cmd.Stdin = os.Stdin
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		cmd.Dir = repo.FullPath
+		cmd.Env = append(os.Environ(), "GHQ_LOOK="+filepath.ToSlash(repo.RelPath))
+		return cmdutil.RunCommand(cmd, true)
+	default:
+		b := &strings.Builder{}
+		b.WriteString("More than one repositories are found; Try more precise name\n")
+		for _, repo := range reposFound {
+			b.WriteString(fmt.Sprintf("       - %s\n", strings.Join(repo.PathParts, "/")))
+		}
+		return errors.New(b.String())
+	}
 }
